fix test build error

+++ tests/src/shell/test_shell_commands.cpp
@@ -20,7 +20,7 @@ const char *system_get_serial_number_string(void) {
 }
 
 static char write_spy_buffer[1024];
-static int write_spy_buffer_index;
+static size_t write_spy_buffer_index;
 static size_t write_spy(const void *data, size_t data_size) {
 	memcpy(write_spy_buffer + write_spy_buffer_index, data, data_size);
 	write_spy_buffer_index += data_size;
